requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
Pygments==2.8.1
pytz==2021.1
sqlparse==0.4.1
typing-extensions==3.7.4.3
dj-static==0.0.6
gunicorn==19.9.0
whitenoise==4.1.2

